Add shift utils to utils
